feat: Запросы пользователей; исправление некоторых ошибок логики

+++ main-service/src/main/java/ru/practicum/dto/EventFullDto.java
@@ -18,7 +18,7 @@ public class EventFullDto {
     private UserShortDto initiator;
     private Location location;
     private Boolean paid;
-    private Integer participantLimit;
+    private Long participantLimit;
     private String publishedOn;
     private String state;
     private String title;
